// server/services/Auth/modal/interface/User.interface.ts
export interface IUser {
    firstName: string;
    lastName: string;
    email: string;
    password: string;
    mobile: string;
    countryCode: string;
    isEmailVerified: boolean;
    isMobileVerified: boolean;
    isActive: boolean;
    createdBy: string;
    updatedBy: string;
    comparePassword(candidatePassword: string): Promise<boolean>;
};